change Long to Integer for durationSeconds

+++ backend/src/main/java/com/nocountry/backend/dto/video/VideoOutputResults.java
@@ -8,7 +8,7 @@ import java.util.List;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record VideoOutputResults(
         String idJob,
-        Long baseVideoDurationSeconds,
+        Integer baseVideoDurationSeconds,
         String state,
         List<VideoOutput> videos,
         LocalDateTime timestamp
